Spawn a new thread for each connection
This implementation is still problematic because it spawns a new thread
for each connection and puts it in a vector. These threads stop once the
client closes the connection but do not get removed from the vector. The
vector is only cleared once the Server object is destroyed, so this is
a memory leak.

A future solution might use a threadpool and epoll().