Fixed bug when binding port 0-1023 & file numbers
Turned out that the signal handling thread wasn't being detached
so when it was destroyed on the server throwing an exception, it was
calling std::terminate.

Also added invalid port checking when assigning to port 0 -- the OS will
not complain about this but for our purposes you cannot bind to port 0.
It's not an actual port, the OS just assigns you to an unused port above
port 1024.

Also start the file numbering from 1 because it seems that's what the
test cases want.